address comments from shuhei
